chore: Add Get command
